Eliminate 'use of old-style cast' for pointers
Eliminate DrawFileReader and merge it into its original base class
DrawReader.  There were no other derived clases of DrawReader, so it does
not need to be virtual.

Add overloads to DrawReader word/byte reading functions to allow a pointer
to an arbitrary type (which must be word sized and aligned) to be passed
without the need for casts.