Fix a crash on exit when doing batch conversion.
Not sure eaxctly what was the root cause of the problem, but
there were two suspicious actions which may have caused an
issue with object lifetime:

 1. If an input file did not exist, DrawView::DrawView() would
    have called deleteLater() on itself and subsequently the
    argument loop in main() did the same.

 2. After batch conversion was finished, whether successful or
    not, exit() was called immediately.  It is not certain whether
    this is the right thing to do when objects pending deletion
    exist;  it may be safer to call QCoreApplication::exit().

With these fixes, there appears to be no crash on exit for either
GUI operation or batch conversion, whether the input file exists
or not.

Fixes #8.